requirements.txt
# Core ML (pinned versions)
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.1.0

# MLOps
dvc==3.30.0
mlflow==2.8.1
boto3==1.28.85

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Monitoring  
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Caching & DB
redis==5.0.1
hiredis==2.2.3
psycopg2-binary==2.9.9

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
